+++ backend/kirikiri/src/main/java/co/kirikiri/domain/ImageContentType.java
@@ -1,4 +1,4 @@
-package co.kirikiri.domain.member;
+package co.kirikiri.domain;
 
 import lombok.Getter;
