+++ backend/kirikiri/src/main/java/co/kirikiri/domain/roadmap/RoadmapNodeImage.java
@@ -1,6 +1,6 @@
 package co.kirikiri.domain.roadmap;
 
-import co.kirikiri.domain.member.ImageContentType;
+import co.kirikiri.domain.ImageContentType;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
