+++ backend/kirikiri/src/main/java/co/kirikiri/service/MemberService.java
@@ -1,8 +1,8 @@
 package co.kirikiri.service;
 
+import co.kirikiri.domain.member.EncryptedPassword;
 import co.kirikiri.domain.member.Member;
 import co.kirikiri.domain.member.MemberProfile;
-import co.kirikiri.domain.member.vo.EncryptedPassword;
 import co.kirikiri.domain.member.vo.Identifier;
 import co.kirikiri.domain.member.vo.Nickname;
 import co.kirikiri.exception.ConflictException;
