+++ backend/kirikiri/src/test/java/co/kirikiri/domain/member/vo/EncryptedPasswordTest.java
@@ -1,5 +1,6 @@
 package co.kirikiri.domain.member.vo;
 
+import co.kirikiri.domain.member.EncryptedPassword;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
