+++ backend/kirikiri/src/test/java/co/kirikiri/persistence/member/MemberProfileRepositoryTest.java
@@ -1,9 +1,9 @@
 package co.kirikiri.persistence.member;
 
+import co.kirikiri.domain.member.EncryptedPassword;
 import co.kirikiri.domain.member.Gender;
 import co.kirikiri.domain.member.Member;
 import co.kirikiri.domain.member.MemberProfile;
-import co.kirikiri.domain.member.vo.EncryptedPassword;
 import co.kirikiri.domain.member.vo.Identifier;
 import co.kirikiri.domain.member.vo.Nickname;
 import co.kirikiri.domain.member.vo.Password;
